Allow vendoring the kubesec.io client.
